+++ test/temmlPostProcess.js
@@ -14,7 +14,7 @@
    * https://mit-license.org/
    */
 
-  const version = "0.6.4";
+  const version = "0.6.5";
 
   function postProcess(block) {
     const labelMap = {};